requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
django-rest-passwordreset==1.3.0
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2023.3
ujson==5.7.0
